# internal/api/supervisor.go
package api

import (
	"context"
	"encoding/json"
	"log"
	"net"
	"net/http"
	_ "net/http/pprof" // registers /debug/pprof handlers on DefaultServeMux
	"sort"
	"strings"
	"sync"
	"time"

	"github.com/gastownhall/gascity/internal/events"
)

// CityInfo describes a managed city for the /v0/cities endpoint.
type CityInfo struct {
	Name            string   `json:"name"`
	Path            string   `json:"path"`
	Running         bool     `json:"running"`
	Status          string   `json:"status,omitempty"`
	Error           string   `json:"error,omitempty"`
	PhasesCompleted []string `json:"phases_completed,omitempty"`
}

// CityResolver provides city lookup for the supervisor API router.
type CityResolver interface {
	// ListCities returns all managed cities with status info.
	ListCities() []CityInfo
	// CityState returns the State for a named city, or nil if not found/not running.
	CityState(name string) State
}

// cachedCityServer pairs a State with its pre-built Server for caching.
type cachedCityServer struct {
	state State
	srv   *Server
}

// SupervisorMux routes API requests to per-city handlers with
// city-namespaced URL paths. It handles:
//   - GET /v0/cities — list managed cities
//   - GET /v0/city/{name} — city detail (status)
//   - /v0/city/{name}/... — route to a specific city's API
//   - /v0/city/{name}/svc/... — route to a specific city's service mount
//   - GET /health — supervisor health
//   - /v0/... (bare) — backward compat, routes to first running city
//   - /svc/... (bare) — route to the sole running city's service mount
type SupervisorMux struct {
	resolver  CityResolver
	readOnly  bool
	version   string
	startedAt time.Time
	server    *http.Server

	// Per-city Server cache. Keyed by city name. Invalidated when
	// the State pointer changes (city restarted → new controllerState).
	cacheMu sync.RWMutex
	cache   map[string]cachedCityServer
}

// NewSupervisorMux creates a SupervisorMux that routes requests to cities
// resolved by the given CityResolver.
func NewSupervisorMux(resolver CityResolver, readOnly bool, version string, startedAt time.Time) *SupervisorMux {
	sm := &SupervisorMux{
		resolver:  resolver,
		readOnly:  readOnly,
		version:   version,
		startedAt: startedAt,
		cache:     make(map[string]cachedCityServer),
	}
	sm.server = &http.Server{Handler: sm.Handler()}
	return sm
}

// Handler returns an http.Handler with the standard middleware chain applied.
func (sm *SupervisorMux) Handler() http.Handler {
	apiInner := withCSRFCheck(http.HandlerFunc(sm.ServeHTTP))
	if sm.readOnly {
		apiInner = withReadOnly(apiInner)
	}
	root := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		if supervisorServicePath(r.URL.Path) {
			// Workspace services apply their own publication and CSRF rules
			// in the per-city server. Do not impose supervisor API policy on
			// top of service mounts.
			sm.ServeHTTP(w, r)
			return
		}
		apiInner.ServeHTTP(w, r)
	})
	// pprof: expose on a separate port for profiling
	go func() {
		_ = http.ListenAndServe("localhost:6060", nil) // default mux has pprof handlers
	}()
	return withLogging(withRecovery(withCORS(root)))
}

// Serve accepts connections on lis. Blocks until stopped.
func (sm *SupervisorMux) Serve(lis net.Listener) error {
	return sm.server.Serve(lis)
}

// Shutdown gracefully shuts down the server.
func (sm *SupervisorMux) Shutdown(ctx context.Context) error {
	return sm.server.Shutdown(ctx)
}

// ServeHTTP dispatches requests to the appropriate city or supervisor-level handler.
func (sm *SupervisorMux) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	path := r.URL.Path

	// Supervisor-level endpoints.
	if path == "/v0/ws" && r.Method == http.MethodGet {
		sm.handleWebSocket(w, r)
		return
	}
	if path == "/v0/cities" && r.Method == http.MethodGet {
		sm.handleCities(w, r)
		return
	}
	if path == "/v0/provider-readiness" && r.Method == http.MethodGet {
		handleProviderReadiness(w, r)
		return
	}
	if path == "/v0/readiness" && r.Method == http.MethodGet {
		handleReadiness(w, r)
		return
	}
	if path == "/health" && r.Method == http.MethodGet {
		sm.handleHealth(w, r)
		return
	}
	// API specs — self-documenting endpoints.
	if path == "/v0/asyncapi.yaml" && r.Method == http.MethodGet {
		handleAsyncAPISpec(w, r)
		return
	}
	if path == "/v0/openapi.yaml" && r.Method == http.MethodGet {
		handleOpenAPISpec(w, r)
		return
	}

	// City creation is supervisor-level.
	if path == "/v0/city" && r.Method == http.MethodPost {
		if sm.readOnly {
			writeError(w, http.StatusForbidden, "read_only", "mutations disabled: server bound to non-localhost address")
			return
		}
		handleCityCreate(w, r)
		return
	}

	// City-namespaced service proxy: /v0/city/{name}/svc/...
	if strings.HasPrefix(path, "/v0/city/") {
		rest := strings.TrimPrefix(path, "/v0/city/")
		idx := strings.IndexByte(rest, '/')
		if idx >= 0 {
			cityName := rest[:idx]
			suffix := rest[idx:]
			if strings.HasPrefix(suffix, "/svc/") {
				sm.serveCityRequest(w, r, cityName, suffix)
				return
			}
		}
		if idx < 0 || rest[:idx] == "" {
			writeError(w, http.StatusBadRequest, "bad_request", "city name required in URL")
			return
		}
	}

	// Bare /svc/... — route to sole running city.
	if strings.HasPrefix(path, "/svc/") {
		cities := sm.resolver.ListCities()
		var running []CityInfo
		for _, c := range cities {
			if c.Running {
				running = append(running, c)
			}
		}
		switch len(running) {
		case 0:
			writeError(w, http.StatusServiceUnavailable, "no_cities", "no cities running")
		case 1:
			sm.serveCityRequest(w, r, running[0].Name, path)
		default:
			writeError(w, http.StatusBadRequest, "city_required",
				"multiple cities running; use /v0/city/{name}/svc/... to specify which city")
		}
		return
	}

	// All other API endpoints are WS-only via GET /v0/ws.
	http.NotFound(w, r)
}

// serveCityRequest resolves a city's State and dispatches to a per-city Server.
func (sm *SupervisorMux) serveCityRequest(w http.ResponseWriter, r *http.Request, cityName, path string) {
	t0 := time.Now()
	state := sm.resolver.CityState(cityName)
	if state == nil {
		sm.cacheMu.Lock()
		delete(sm.cache, cityName)
		sm.cacheMu.Unlock()
		writeError(w, http.StatusNotFound, "not_found", "city not found or not running: "+cityName)
		return
	}
	t1 := time.Now()

	srv := sm.getCityServer(cityName, state)
	t2 := time.Now()

	r2 := r.Clone(r.Context())
	r2.URL.Path = path
	r2.URL.RawPath = ""
	srv.mux.ServeHTTP(w, r2)
	t3 := time.Now()

	total := t3.Sub(t0)
	if total > 500*time.Millisecond {
		log.Printf("SLOW serveCityRequest %s: resolve=%s getServer=%s handler=%s total=%s",
			path, t1.Sub(t0), t2.Sub(t1), t3.Sub(t2), total)
	}
}

// getCityServer returns a cached per-city Server, creating one if the
// cache is empty or the State pointer changed (city was restarted).
func (sm *SupervisorMux) getCityServer(name string, state State) *Server {
	sm.cacheMu.RLock()
	if cached, ok := sm.cache[name]; ok && cached.state == state {
		sm.cacheMu.RUnlock()
		return cached.srv
	}
	sm.cacheMu.RUnlock()

	srv := New(state)
	if sm.readOnly {
		srv = NewReadOnly(state)
	}

	sm.cacheMu.Lock()
	sm.cache[name] = cachedCityServer{state: state, srv: srv}
	sm.cacheMu.Unlock()

	return srv
}

func supervisorServicePath(path string) bool {
	if strings.HasPrefix(path, "/svc/") {
		return true
	}
	if !strings.HasPrefix(path, "/v0/city/") {
		return false
	}
	rest := strings.TrimPrefix(path, "/v0/city/")
	idx := strings.IndexByte(rest, '/')
	if idx < 0 {
		return false
	}
	return strings.HasPrefix(rest[idx:], "/svc/")
}

func (sm *SupervisorMux) handleCities(w http.ResponseWriter, _ *http.Request) {
	writeJSON(w, http.StatusOK, sm.citiesList())
}

// handleGlobalEventStream streams SSE events from all running cities,
// tagged with city name. The cursor format for reconnection is
// "city1:seq1,city2:seq2" via Last-Event-ID or ?after_cursor.
func (sm *SupervisorMux) handleGlobalEventStream(w http.ResponseWriter, r *http.Request) {
	mux := sm.buildMultiplexer()

	// Parse cursor from Last-Event-ID or query param.
	cursor := r.Header.Get("Last-Event-ID")
	if cursor == "" {
		cursor = r.URL.Query().Get("after_cursor")
	}
	cursors := events.ParseCursor(cursor)

	mw, err := mux.Watch(r.Context(), cursors)
	if err != nil {
		writeError(w, http.StatusServiceUnavailable, "internal", "failed to start global event watcher: "+err.Error())
		return
	}

	w.Header().Set("Content-Type", "text/event-stream")
	w.Header().Set("Cache-Control", "no-cache")
	w.Header().Set("Connection", "keep-alive")
	w.WriteHeader(http.StatusOK)
	if err := http.NewResponseController(w).Flush(); err != nil {
		_ = err
	}

	// Stream tagged events with composite cursor IDs. We use a
	// dedicated loop because the SSE id must be a composite per-city
	// cursor, not a scalar Seq.
	streamProjectedGlobalEvents(r.Context(), w, mw, cursors, sm.resolver)
}

// handleGlobalEventList returns events from all running cities, sorted
// by timestamp, with each event tagged with its source city.
func (sm *SupervisorMux) handleGlobalEventList(w http.ResponseWriter, r *http.Request) {
	mux := sm.buildMultiplexer()

	q := r.URL.Query()
	filter := events.Filter{
		Type:  q.Get("type"),
		Actor: q.Get("actor"),
	}
	if v := q.Get("since"); v != "" {
		if d, err := time.ParseDuration(v); err == nil {
			filter.Since = time.Now().Add(-d)
		}
	}

	evts, err := mux.ListAll(filter)
	if err != nil {
		writeError(w, http.StatusInternalServerError, "internal", err.Error())
		return
	}
	if evts == nil {
		evts = []events.TaggedEvent{}
	}
	writeJSON(w, http.StatusOK, listResponse{Items: evts, Total: len(evts)})
}

// buildMultiplexer creates a Multiplexer from all running cities'
// event providers.
func (sm *SupervisorMux) buildMultiplexer() *events.Multiplexer {
	mux := events.NewMultiplexer()
	cities := sm.resolver.ListCities()
	for _, c := range cities {
		if !c.Running {
			continue
		}
		state := sm.resolver.CityState(c.Name)
		if state == nil {
			continue
		}
		ep := state.EventProvider()
		if ep == nil {
			continue
		}
		mux.Add(c.Name, ep)
	}
	return mux
}

// globalEventList returns events aggregated from all running cities.
func (sm *SupervisorMux) globalEventList(req *socketRequestEnvelope) (any, error) {
	mux := sm.buildMultiplexer()
	var payload socketEventsListPayload
	if len(req.Payload) > 0 {
		_ = json.Unmarshal(req.Payload, &payload)
	}
	filter := events.Filter{Type: payload.Type, Actor: payload.Actor}
	if payload.Since != "" {
		if d, err := time.ParseDuration(payload.Since); err == nil {
			filter.Since = time.Now().Add(-d)
		}
	}
	evts, err := mux.ListAll(filter)
	if err != nil {
		return nil, err
	}
	if evts == nil {
		evts = []events.TaggedEvent{}
	}
	return listResponse{Items: evts, Total: len(evts)}, nil
}

func (sm *SupervisorMux) handleHealth(w http.ResponseWriter, _ *http.Request) {
	writeJSON(w, http.StatusOK, sm.healthResponse())
}

func (sm *SupervisorMux) citiesList() listResponse {
	cities := sm.resolver.ListCities()
	sort.Slice(cities, func(i, j int) bool { return cities[i].Name < cities[j].Name })
	return listResponse{Items: cities, Total: len(cities)}
}

func (sm *SupervisorMux) healthResponse() map[string]any {
	cities := sm.resolver.ListCities()
	var running int
	// Use the first city for startup info (single-city deployments).
	var startup map[string]any
	for _, c := range cities {
		if c.Running {
			running++
		}
		if startup == nil {
			if c.Running {
				startup = map[string]any{
					"ready":            true,
					"phase":            "running",
					"phases_completed": allStartupPhases(),
				}
			} else {
				startup = map[string]any{
					"ready":            false,
					"phase":            c.Status,
					"phases_completed": c.PhasesCompleted,
				}
			}
		}
	}
	resp := map[string]any{
		"status":         "ok",
		"version":        sm.version,
		"uptime_sec":     int(time.Since(sm.startedAt).Seconds()),
		"cities_total":   len(cities),
		"cities_running": running,
	}
	if startup != nil {
		resp["startup"] = startup
	}
	return resp
}

// allStartupPhases returns the ordered list of all startup phases.
func allStartupPhases() []string {
	return []string{
		"loading_config",
		"starting_bead_store",
		"resolving_formulas",
		"adopting_sessions",
		"starting_agents",
	}
}